Fix "no bin target" error and add a test
The build error was:

    error: no bin target named `pgrx_embed_clerk_fdw`.

This is a side-effect of pgrx v0.12.x; pgcentralfoundation/pgrx#1888
documents the fix: adding a no-op embed target in
`src/bin/pgrx_embed_clerk_fdw.rs` and setting `lib.crate-type` to:

    crate-type = ["cdylib", "lib"]

Hopefully this funky workaround can be removed in the future.

The test simply ensures that `CREATE EXTENSION` works; the update to
`extension_ci.yml` ensures it passes on all supported Postgres versions.